refactor: move all ui primitives to @ajh/ui, desktop re-exports them
- add MarkdownMessage, StreamingText, OptionTile to packages/ui
- upgrade Toast in packages/ui to provider pattern (ToastProvider/useToast)
- update packages/ui index.ts with all new exports
- replace every component in apps/desktop/components/ui with thin re-exports from @ajh/ui
- keep UpdateBanner in desktop (app-specific: IPC/updater dependency)
- fix Toast storybook to use new provider API
- re-add updater preload bridge and use-updater service hook
