+++ apps/desktop/src/renderer/components/ui/EmptyState.tsx
@@ -1,31 +1,1 @@
-import type { LucideIcon } from 'lucide-react';
-import type { ReactNode } from 'react';
-import { cn } from '@/lib/cn';
-
-interface EmptyStateProps {
-  icon: LucideIcon;
-  title: string;
-  description?: string;
-  action?: ReactNode;
-  className?: string;
-}
-
-/**
- * Consistent empty-content placeholder used across all feature pages.
- */
-export function EmptyState({ icon: Icon, title, description, action, className }: EmptyStateProps) {
-  return (
-    <div
-      className={cn('flex flex-col items-center justify-center gap-4 py-16 text-center', className)}
-    >
-      <div className="flex h-14 w-14 items-center justify-center rounded-2xl bg-white/5 ring-1 ring-white/[0.08]">
-        <Icon size={24} className="text-foreground/30" />
-      </div>
-      <div className="space-y-1">
-        <p className="text-sm font-medium text-foreground/60">{title}</p>
-        {description && <p className="max-w-xs text-xs text-foreground/35">{description}</p>}
-      </div>
-      {action && <div className="mt-2">{action}</div>}
-    </div>
-  );
-}
+export { EmptyState } from '@ajh/ui';
