+++ apps/desktop/src/renderer/components/ui/TextArea.tsx
@@ -1,21 +1,1 @@
-import type { TextareaHTMLAttributes } from 'react';
-import { cn } from '@/lib/cn';
-
-export interface TextAreaProps extends TextareaHTMLAttributes<HTMLTextAreaElement> {
-  variant?: 'default' | 'glass';
-}
-
-export function TextArea({ className, variant = 'default', ...props }: TextAreaProps) {
-  return (
-    <textarea
-      className={cn(
-        'input-field w-full resize-none text-sm leading-relaxed text-foreground placeholder:text-foreground/30',
-        variant === 'glass' &&
-          'rounded-lg px-3 py-2 border border-white/[0.06] bg-[rgba(15,15,25,0.45)] hover:border-white/10 focus:border-brand/35',
-        variant === 'default' && 'bg-transparent',
-        className
-      )}
-      {...props}
-    />
-  );
-}
+export { TextArea, type TextAreaProps } from '@ajh/ui';
